:recycle: import required testing libraries on per-spec file basis
The practice of using a setup file is eluding me. Just putting a setup.specs.js in the parent level directory and doing all the imports does not provide global access to future spec files to things like expect from chai or React or the Adapter or shallow from enzyme.